saveManager for TB
