+++ src/scripts/shared/util/config-options.ts
@@ -1,0 +1,3 @@
+import { ref } from "vue"
+
+export const gridSize = ref(20)
